pyproject.toml
[project]
name = "xsliv-downloader"
version = "0.4.0"
description = "Парсер и установщик фото/видео с xxsliv.com"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fake-useragent>=2.2.0",
    "pillow>=11.3.0",
    "questionary>=2.1.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
